vinyl: fix check vinyl_dir existence at bootstrap
During implementation of openSUSE got failed box-tap/cfg.test.lua test.
Found that when memtx_dir didn't exist and vinyl_dir existed and also
errno was set to ENOENT, box configuration succeeded, but it shouldn't.
Reason of this wrong behaviour was that not all of the failure paths in
xdir_scan() set errno, but the caller assumed it.

Debugging src/box/xlog.c found that all checks were correct, but at:

  src/box/vy_log.c:vy_log_bootstrap()
  src/box/vy_log.c:vy_log_begin_recovery()

the checks on of the errno on ENOENT blocked the negative return from:

  src/box/xlog.c:xdir_scan()

Found that errno was already set to ENOENT before the xdir_scan() call.
To fix the issue the errno could be clean before the call to xdir_scan,
because we are interesting in it only from xdir_scan function.

After discussions found that there were alternative better solution to
fix it. The fix with resetting errno was not good because xdir_scan()
was not system call in real and some internal routines could set it
to ENOENT itself, so it couldn't be controled from outside of function.

To be sure in behaviour of the changing errno decided to pass a flag to
xdir_scan() if the directory should exist.

Closes #4594
Needed for #4562

Co-authored-by: Alexander Turenko <alexander.turenko@tarantool.org>